[SPARK-43292][CORE][CONNECT] Move `ExecutorClassLoader` to `core` module and simplify `Executor#addReplClassLoaderIfNeeded`
### What changes were proposed in this pull request?
This pr move `ExecutorClassLoader` from `repl` module to `core` module an put it into `executor` package, then `ArtifactManagerSuite` can test using maven.

On the other hand, this pr removed reflection calls in the `Executor#addReplClassLoaderIfNeeded` due to `ExecutorClassLoader` and `Executor` are in the same module after this pr.

### Why are the changes needed?
1. `ExecutorClassLoader` only be used by `Executor`, it is more suitable for placing in the `core` module
2. Make `ArtifactManagerSuite` can test using maven.

### Does this PR introduce _any_ user-facing change?
No, just for maven test

### How was this patch tested?
- Pass GitHub Actions
- Manual test

Run the following commands

```
build/mvn clean install -DskipTests -Phive
build/mvn test -pl connector/connect/server
```

**Before**

`ArtifactManagerSuite` test failed due to:

```
23/04/26 16:36:57.140 ScalaTest-main-running-DiscoverySuite ERROR Executor: Could not find org.apache.spark.repl.ExecutorClassLoader on classpath!
```

**After**

All tests passed.

```
Run completed in 10 seconds, 494 milliseconds.
Total number of tests run: 560
Suites: completed 11, aborted 0
Tests: succeeded 560, failed 0, canceled 0, ignored 0, pending 0
All tests passed.
```

Closes #40956 from LuciferYang/SPARK-43292.

Lead-authored-by: yangjie01 <yangjie01@baidu.com>
